# expression: implement vectorized evaluation for `builtinLocateBinary3ArgsSig`