Signed-off-by: baishen <baishen2009@gmail.com>

<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for builtinLocateBinary3ArgsSig.
[12106](https://github.com/pingcap/tidb/issues/12106)

### What is changed and how it works?
```
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary3ArgsSig-VecBuiltinFunc-4         	   85585	     13222 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary3ArgsSig-NonVecBuiltinFunc-4      	   19460	     78087 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary3ArgsSig-VecBuiltinFunc#01-4      	   62563	     16759 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary3ArgsSig-NonVecBuiltinFunc#01-4   	   10000	    131020 ns/op	       0 B/op	       0 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test